chore: refactored file stricture 1/2
