+++ app/components/Views/TrendingView/sections.config.tsx
@@ -1,30 +1,30 @@
 import React, { PropsWithChildren } from 'react';
 import type { NavigationProp, ParamListBase } from '@react-navigation/native';
 import type { TrendingAsset } from '@metamask/assets-controllers';
-import Routes from '../../../../constants/navigation/Routes';
-import { strings } from '../../../../../locales/i18n';
-import TrendingTokenRowItem from '../../../UI/Trending/components/TrendingTokenRowItem/TrendingTokenRowItem';
-import TrendingTokensSkeleton from '../../../UI/Trending/components/TrendingTokenSkeleton/TrendingTokensSkeleton';
-import PerpsMarketRowItem from '../../../UI/Perps/components/PerpsMarketRowItem';
-import type { PerpsMarketData } from '../../../UI/Perps/controllers/types';
-import PredictMarket from '../../../UI/Predict/components/PredictMarket';
-import type { PredictMarket as PredictMarketType } from '../../../UI/Predict/types';
-import type { PerpsNavigationParamList } from '../../../UI/Perps/types/navigation';
-import PredictMarketSkeleton from '../../../UI/Predict/components/PredictMarketSkeleton';
-import { usePredictMarketData } from '../../../UI/Predict/hooks/usePredictMarketData';
-import { usePerpsMarkets } from '../../../UI/Perps/hooks';
-import { PerpsConnectionProvider } from '../../../UI/Perps/providers/PerpsConnectionProvider';
-import { PerpsStreamProvider } from '../../../UI/Perps/providers/PerpsStreamManager';
+import Routes from '../../../constants/navigation/Routes';
+import { strings } from '../../../../locales/i18n';
+import TrendingTokenRowItem from '../../UI/Trending/components/TrendingTokenRowItem/TrendingTokenRowItem';
+import TrendingTokensSkeleton from '../../UI/Trending/components/TrendingTokenSkeleton/TrendingTokensSkeleton';
+import PerpsMarketRowItem from '../../UI/Perps/components/PerpsMarketRowItem';
+import type { PerpsMarketData } from '../../UI/Perps/controllers/types';
+import PredictMarket from '../../UI/Predict/components/PredictMarket';
+import type { PredictMarket as PredictMarketType } from '../../UI/Predict/types';
+import type { PerpsNavigationParamList } from '../../UI/Perps/types/navigation';
+import PredictMarketSkeleton from '../../UI/Predict/components/PredictMarketSkeleton';
+import { usePredictMarketData } from '../../UI/Predict/hooks/usePredictMarketData';
+import { usePerpsMarkets } from '../../UI/Perps/hooks';
+import { PerpsConnectionProvider } from '../../UI/Perps/providers/PerpsConnectionProvider';
+import { PerpsStreamProvider } from '../../UI/Perps/providers/PerpsStreamManager';
 import { Box, IconName } from '@metamask/design-system-react-native';
-import type { SiteData } from '../../../UI/Sites/components/SiteRowItem/SiteRowItem';
-import SiteRowItemWrapper from '../../../UI/Sites/components/SiteRowItemWrapper/SiteRowItemWrapper';
-import SiteSkeleton from '../../../UI/Sites/components/SiteSkeleton/SiteSkeleton';
-import { useSitesData } from '../../../UI/Sites/hooks/useSiteData/useSitesData';
-import { useTrendingSearch } from '../../../UI/Trending/hooks/useTrendingSearch/useTrendingSearch';
-import { filterMarketsByQuery } from '../../../UI/Perps/utils/marketUtils';
-import PredictMarketRowItem from '../../../UI/Predict/components/PredictMarketRowItem';
-import SectionCard from '../components/SectionCard/SectionCard';
-import SectionCarrousel from '../components/SectionCarrousel/SectionCarrousel';
+import type { SiteData } from '../../UI/Sites/components/SiteRowItem/SiteRowItem';
+import SiteRowItemWrapper from '../../UI/Sites/components/SiteRowItemWrapper/SiteRowItemWrapper';
+import SiteSkeleton from '../../UI/Sites/components/SiteSkeleton/SiteSkeleton';
+import { useSitesData } from '../../UI/Sites/hooks/useSiteData/useSitesData';
+import { useTrendingSearch } from '../../UI/Trending/hooks/useTrendingSearch/useTrendingSearch';
+import { filterMarketsByQuery } from '../../UI/Perps/utils/marketUtils';
+import PredictMarketRowItem from '../../UI/Predict/components/PredictMarketRowItem';
+import SectionCard from './components/Sections/SectionTypes/SectionCard';
+import SectionCarrousel from './components/Sections/SectionTypes/SectionCarrousel';
 
 export type SectionId = 'predictions' | 'tokens' | 'perps' | 'sites';
